Move substate-cli/state to state
This PR resolve duplicated "state" package by moving state files in
substate-cli/ to state/. Memory statedb is extended to cover functions
of Aida stateDB interface.
